refactor:generelise DatabaseQueryLog to QueryLog and move it in datastudio app
