+++ backend/hexa/data_studio/models.py
@@ -80,3 +80,73 @@ class SavedQuery(Base):
             raise PermissionDenied
 
         return self.delete()
+
+
+class QueryLogQuerySet(BaseQuerySet):
+    def filter_for_user(self, user: AnonymousUser | UserInterface) -> models.QuerySet:
+        return self._filter_for_user_and_query_object(
+            user,
+            models.Q(workspace__in=Workspace.objects.filter_for_user(user)),
+        )
+
+
+class QueryLog(Base):
+    """Audit log entry for a query executed against a data source via the API."""
+
+    # SQLSTATE code reported for a successfully completed statement
+    # (https://www.postgresql.org/docs/current/errcodes-appendix.html)
+    SQLSTATE_SUCCESS = "00000"
+
+    class Status(models.TextChoices):
+        SUCCESS = "SUCCESS"
+        # The query reached the data source but failed (invalid SQL, timeout, ...)
+        ERROR = "ERROR"
+        # The user was not allowed to run queries against the target
+        DENIED = "DENIED"
+        # The query was rejected before reaching the data source (e.g. multiple statements)
+        REJECTED = "REJECTED"
+
+    class Origin(models.TextChoices):
+        # The client did not identify itself; every query arrives through the API anyway
+        OTHER = "OTHER"
+        DATA_STUDIO = "DATA_STUDIO"
+
+    workspace = models.ForeignKey(
+        Workspace, on_delete=models.CASCADE, related_name="query_logs"
+    )
+    user = models.ForeignKey(
+        User,
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name="query_logs",
+    )
+    query = models.TextField()
+    status = models.CharField(max_length=10, choices=Status.choices)
+    # SQLSTATE error code (https://www.postgresql.org/docs/current/errcodes-appendix.html);
+    # SQLSTATE_SUCCESS on success, null when the query never reached the data source
+    result_code = models.CharField(max_length=5, null=True, blank=True)
+    error_message = models.TextField(null=True, blank=True)
+    duration_ms = models.PositiveIntegerField(null=True)
+    row_count = models.PositiveIntegerField(null=True)
+    truncated = models.BooleanField(default=False)
+    origin = models.CharField(
+        max_length=20, choices=Origin.choices, default=Origin.OTHER
+    )
+    # Free-text identifier of what was queried (e.g. "workspace_database"),
+    # kept as plain text rather than choices so new source types don't need a migration.
+    target = models.TextField()
+
+    objects = QueryLogQuerySet.as_manager()
+
+    class Meta:
+        ordering = ["-created_at"]
+        indexes = [
+            models.Index(fields=["workspace", "-created_at"]),
+            models.Index(fields=["user", "-created_at"]),
+            models.Index(fields=["status"]),
+            models.Index(fields=["origin"]),
+            models.Index(fields=["target"]),
+        ]
+
+    def __str__(self) -> str:
+        return f"{self.workspace.slug} - {self.status} - {self.created_at:%Y-%m-%d %H:%M:%S}"
