+++ backend/hexa/data_studio/tests/test_models.py
@@ -1,7 +1,18 @@
+from unittest.mock import MagicMock
+
+from django.contrib.auth.models import AnonymousUser
 from django.core.exceptions import PermissionDenied
 
 from hexa.core.test import TestCase
-from hexa.data_studio.models import SavedQuery
+from hexa.data_studio.models import QueryLog, SavedQuery
+from hexa.pipelines.authentication import PipelineRunUser
+from hexa.pipelines.models import Pipeline, PipelineRun
+from hexa.user_management.models import User
+from hexa.workspaces.models import (
+    Workspace,
+    WorkspaceMembership,
+    WorkspaceMembershipRole,
+)
 
 from .testutils import SavedQueryTestMixin
 
@@ -82,3 +93,100 @@ class SavedQueryModelTest(SavedQueryTestMixin, TestCase):
         query = self._create(user=self.USER_EDITOR)
         with self.assertRaises(PermissionDenied):
             query.delete_if_has_perm(principal=self.USER_VIEWER)
+
+
+class QueryLogModelTest(TestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.USER_SUPERUSER = User.objects.create_user(
+            "superuser@bluesquarehub.com", "superuserpassword", is_superuser=True
+        )
+        cls.USER_MEMBER = User.objects.create_user(
+            "member@bluesquarehub.com", "memberpassword"
+        )
+        cls.USER_OUTSIDER = User.objects.create_user(
+            "outsider@bluesquarehub.com", "outsiderpassword"
+        )
+        cls.WORKSPACE_1 = Workspace.objects.create_if_has_perm(
+            cls.USER_SUPERUSER,
+            name="Workspace 1",
+            description="First workspace",
+            countries=[],
+        )
+        cls.WORKSPACE_2 = Workspace.objects.create_if_has_perm(
+            cls.USER_SUPERUSER,
+            name="Workspace 2",
+            description="Second workspace",
+            countries=[],
+        )
+        WorkspaceMembership.objects.create(
+            user=cls.USER_MEMBER,
+            workspace=cls.WORKSPACE_1,
+            role=WorkspaceMembershipRole.VIEWER,
+        )
+        cls.LOG_WS1_MEMBER = QueryLog.objects.create(
+            workspace=cls.WORKSPACE_1,
+            user=cls.USER_MEMBER,
+            query="SELECT 1",
+            status=QueryLog.Status.SUCCESS,
+            result_code="00000",
+            target="workspace_database",
+        )
+        cls.LOG_WS1_OTHER_USER = QueryLog.objects.create(
+            workspace=cls.WORKSPACE_1,
+            user=cls.USER_SUPERUSER,
+            query="SELECT 2",
+            status=QueryLog.Status.ERROR,
+            result_code="42601",
+            target="workspace_database",
+        )
+        cls.LOG_WS2 = QueryLog.objects.create(
+            workspace=cls.WORKSPACE_2,
+            user=cls.USER_SUPERUSER,
+            query="SELECT 3",
+            status=QueryLog.Status.SUCCESS,
+            result_code="00000",
+            target="workspace_database",
+        )
+
+    def test_filter_for_user_member_sees_own_workspace_logs(self):
+        # Members see all logs of their workspaces, not only their own queries
+        self.assertEqual(
+            {self.LOG_WS1_MEMBER, self.LOG_WS1_OTHER_USER},
+            set(QueryLog.objects.filter_for_user(self.USER_MEMBER)),
+        )
+
+    def test_filter_for_user_outsider_sees_nothing(self):
+        self.assertEqual(
+            0, QueryLog.objects.filter_for_user(self.USER_OUTSIDER).count()
+        )
+
+    def test_filter_for_user_anonymous_sees_nothing(self):
+        self.assertEqual(0, QueryLog.objects.filter_for_user(AnonymousUser()).count())
+
+    def test_filter_for_user_superuser_sees_everything(self):
+        self.assertEqual(
+            3, QueryLog.objects.filter_for_user(self.USER_SUPERUSER).count()
+        )
+
+    def test_filter_for_user_pipeline_run_user_sees_its_workspace_logs(self):
+        pipeline_run = MagicMock(PipelineRun)
+        pipeline_run.pipeline = MagicMock(Pipeline)
+        pipeline_run.pipeline.workspace = self.WORKSPACE_1
+        pipeline_run.pipeline.workspace_id = self.WORKSPACE_1.id
+        pipeline_user = PipelineRunUser(pipeline_run)
+
+        self.assertEqual(
+            {self.LOG_WS1_MEMBER, self.LOG_WS1_OTHER_USER},
+            set(QueryLog.objects.filter_for_user(pipeline_user)),
+        )
+
+    def test_log_survives_user_deletion(self):
+        # The audit trail must be kept even after the author is deleted
+        self.USER_MEMBER.delete()
+        self.LOG_WS1_MEMBER.refresh_from_db()
+        self.assertIsNone(self.LOG_WS1_MEMBER.user)
+
+    def test_workspace_deletion_removes_logs(self):
+        self.WORKSPACE_2.delete()
+        self.assertFalse(QueryLog.objects.filter(id=self.LOG_WS2.id).exists())
